+++ src/models/BnnModel.py
@@ -1,5 +1,8 @@
+"""
+This module contains the implementation of an improved Bayesian Neural Network (BNN) model
+"""
 import torch
-import torch.nn as nn
+from torch import nn
 from torch.distributions import Normal
 from torch.nn import functional as F
 
@@ -11,6 +14,8 @@ class BnnLayer(nn.Module):
     Args:
         in_features (int): Number of input features.
         out_features (int): Number of output features.
+        batch_norm (bool): Flag to enable batch normalization.
+
 
     Attributes:
         in_features (int): Number of input features.
@@ -19,6 +24,7 @@ class BnnLayer(nn.Module):
         weight_std (nn.Parameter): Standard deviation parameters for weight distribution.
         bias_mu (nn.Parameter): Mean parameters for bias distribution.
         bias_std (nn.Parameter): Standard deviation parameters for bias distribution.
+        batch_norm (bool): Flag indicating whether batch normalization is used.
 
     Methods:
         reset_parameters(): Initialize parameters with specific initialization schemes.
@@ -26,11 +32,12 @@ class BnnLayer(nn.Module):
 
     """
 
-    def __init__(self, in_features, out_features):
-        super(BnnLayer, self).__init__()
+    def __init__(self, in_features, out_features, batch_norm=False):
+        super().__init__()
 
         self.in_features = in_features
         self.out_features = out_features
+        self.batch_norm = batch_norm
 
         # Parameters for the weight distribution
         self.weight_mu = nn.Parameter(torch.Tensor(out_features, in_features))
@@ -39,6 +46,10 @@ class BnnLayer(nn.Module):
         # Parameters for the bias distribution
         self.bias_mu = nn.Parameter(torch.Tensor(out_features))
         self.bias_std = nn.Parameter(torch.Tensor(out_features))
+
+        if self.batch_norm:
+            # Parameters for batch normalization
+            self.bn = nn.BatchNorm1d(out_features)
 
         # Initialize parameters
         self.reset_parameters()
@@ -52,10 +63,10 @@ class BnnLayer(nn.Module):
         parameters are initialized with a constant value.
         """
         nn.init.kaiming_normal_(self.weight_mu, mode="fan_in", nonlinearity="relu")
-        nn.init.constant_(self.weight_std, -5.0)
+        nn.init.constant_(self.weight_std, -4.0)
 
         nn.init.zeros_(self.bias_mu)
-        nn.init.constant_(self.bias_std, -5.0)
+        nn.init.constant_(self.bias_std, -4.0)
 
     def forward(self, x):
         """
@@ -73,7 +84,13 @@ class BnnLayer(nn.Module):
         bias = Normal(self.bias_mu, torch.exp(self.bias_std)).rsample()
 
         # Apply linear transformation to input using sampled weights and biases
-        return F.linear(x, weight, bias)
+        x = F.linear(x, weight, bias)
+
+        if self.batch_norm:
+            # Apply batch normalization
+            x = self.bn(x)
+
+        return x
 
 
 class BayesianModel(nn.Module):
@@ -94,12 +111,13 @@ class BayesianModel(nn.Module):
 
     """
 
-    def __init__(self, input_size, hidden_size, output_size):
-        super(BayesianModel, self).__init__()
+    def __init__(self, input_size, hidden_size1, hidden_size2, output_size):
+        super().__init__()
 
         # Define the architecture of the Bayesian neural network
-        self.layer1 = BnnLayer(input_size, hidden_size)
-        self.layer2 = BnnLayer(hidden_size, output_size)
+        self.layer1 = BnnLayer(input_size, hidden_size1)
+        self.layer2 = BnnLayer(hidden_size1, hidden_size2)
+        self.layer3 = BnnLayer(hidden_size2, output_size)
 
     def forward(self, x):
         """
@@ -113,7 +131,23 @@ class BayesianModel(nn.Module):
 
         """
         # Apply ReLU activation to the output of the first layer
-        x = F.relu(self.layer1(x))
+        x = F.leaky_relu(self.layer1(x))
         # Pass the result through the second layer
-        x = self.layer2(x)
+        x = F.leaky_relu(self.layer2(x))
+        # Pass the result through the third layer
+        x = self.layer3(x)
         return x
+
+    def predict(self, x):
+        """
+        Desc:
+            Sample predictions from the posterior distribution of the model.
+        Args:
+            x (array-like): Input features.
+        Returns:
+            tensor: Predicted values.
+        """
+        self.eval()
+        with torch.no_grad():
+            predictions = self.forward(x)
+        return predictions
